+++ crates/ludtwig/CHANGELOG.md
@@ -1,5 +1,8 @@
 # NEXT-VERSION
 
+- [#143](https://github.com/MalteJanz/ludtwig/issues/143) Raised minimum required Rust version to 1.85 and upgraded used
+  rust edition to 2024
+
 # v0.10.0
 
 - Bumped version of `ludtwig-parser` to `0.7.0`, which includes more parsing support (e.g. arrow functions, symfony